Wrapped ResetPassword client in Suspense
